% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{assessRSqlDf}
\alias{assessRSqlDf}
\title{assessRSqlDf}
\usage{
assessRSqlDf(values.df)
}
\arguments{
\item{values.df}{Data Frame with corresponding values and fields as colnames}
}
\description{
This function prepares data frame to be inserted in a db
}
\author{
ken4rab
}
