% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{binary_operator_to_sql}
\alias{binary_operator_to_sql}
\title{Format a binary operator with name \code{x}}
\usage{
binary_operator_to_sql(x)
}
\arguments{
\item{x}{operator name}
}
\description{
Function created converts \code{e1} and \code{e2} to SQL using
\code{to_sql} before formatting them on either side of operator
name \code{x}.
}
