% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsql_statement.R
\docType{methods}
\name{$,rsql_select_statement-method}
\alias{$,rsql_select_statement-method}
\title{Overloading the $ operator for access to column references}
\usage{
\S4method{$}{rsql_select_statement}(x, name)
}
\description{
Overloading the $ operator for access to column references
}
