% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsql_table.R
\docType{methods}
\name{$,rsql_table-method}
\alias{$,rsql_table-method}
\title{Overloading the $ operator for access to column references}
\usage{
\S4method{$}{rsql_table}(x, name)
}
\description{
Overloading the $ operator for access to column references
}
