% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{createRSQL}
\alias{createRSQL}
\title{Produces a RSQL object}
\usage{
createRSQL(drv, dbname, user = NULL, password = NULL, host = NULL, port = NULL)
}
\arguments{
\item{drv}{Driver name}

\item{dbname}{Database name}

\item{user}{Database user name}

\item{password}{Database password}

\item{host}{Database host}

\item{port}{Database port}
}
\description{
Produces a RSQL object
}
