% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{dequote}
\alias{dequote}
\title{Removes the quotes from the string}
\usage{
dequote(text)
}
\arguments{
\item{text}{The string to remove the quotes from.}
}
\description{
Removes the quotes from the string
}
\author{
ken4rab
}
