% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{df_verify}
\alias{df_verify}
\title{Checks that the columns are in the data.frame}
\usage{
df_verify(dataframe, columns)
}
\arguments{
\item{dataframe}{The data.frame}

\item{columns}{The columns to check}
}
\description{
Checks that the columns are in the data.frame
}
