% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lib.R
\name{execute_get_insert}
\alias{execute_get_insert}
\title{execute_get_insert}
\usage{
execute_get_insert(id_procesamiento, sql_select, sql_insert, ...)
}
\arguments{
\item{id_procesamiento}{TEST}

\item{sql_select}{TEST}

\item{sql_insert}{TEST}
}
\description{
execute_get_insert
}
