% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{function_to_sql}
\alias{function_to_sql}
\title{Format a regular function with name \code{x}}
\usage{
function_to_sql(x)
}
\arguments{
\item{x}{function name}
}
\description{
Function created converts arguments \code{...} to SQL using
\code{to_sql} before wrapping them in parenthesis and 
prepending function name \code{x}.
}
