% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{getMtcarsdbPath}
\alias{getMtcarsdbPath}
\title{getCarsdbPath}
\usage{
getMtcarsdbPath(copy = TRUE)
}
\arguments{
\item{copy}{a boolean that states whether it should be copied to the home directory or not.}
}
\description{
getCarsdbPath
}
\author{
ken4rab
}
