% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{\%IN\%}
\alias{\%IN\%}
\title{Operator IN for multiple columns}
\usage{
x \%IN\% y
}
\arguments{
\item{x}{vector x}

\item{y}{vector y}
}
\description{
Operator IN for multiple columns
}
\author{
ken4rab
}
