% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{is_quoted}
\alias{is_quoted}
\title{Determines if the string is quoted or not}
\usage{
is_quoted(text, quotes_symbols = c("'", "'"))
}
\arguments{
\item{text}{The text to test}

\item{quotes_symbols}{The quotation characters}
}
\description{
Determines if the string is quoted or not
}
