% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-lib.R
\name{loggerSetupFile}
\alias{loggerSetupFile}
\title{loggerSetupFile}
\usage{
loggerSetupFile(log.file, default.threshold = "info")
}
\arguments{
\item{log.file}{filename to log on}

\item{default.threshold}{default threshold for logger}
}
\description{
loggerSetupFile
}
\author{
kenarab
}
