% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{needs_quotes}
\alias{needs_quotes}
\title{Determines string type which needs quotes in an SQL statement}
\usage{
needs_quotes(text)
}
\arguments{
\item{text}{The text to test}
}
\description{
Determines string type which needs quotes in an SQL statement
}
\author{
ken4rab
}
