% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{parse_where_clause}
\alias{parse_where_clause}
\title{Parses a where clause.}
\usage{
parse_where_clause(where_clause_list = c())
}
\arguments{
\item{where_clause_list}{The list of params}
}
\description{
Parses a where clause.
}
\author{
ken4rab
}
