% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsql_statement.R
\name{+.rsql_select_statement}
\alias{+.rsql_select_statement}
\title{Function for merging SELECT statements with additional elements}
\usage{
\method{+}{rsql_select_statement}(e1, e2)
}
\description{
In most cases, this should not be used directly by users.
}
