% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{re_quote}
\alias{re_quote}
\title{TODO: WHAT IS THIS FUNCTION DOING AGAIN?}
\usage{
re_quote(text, quotes = "'")
}
\arguments{
\item{text}{The string}

\item{quotes}{The quotes}
}
\description{
TODO: WHAT IS THIS FUNCTION DOING AGAIN?
}
