% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{remove_quotes}
\alias{remove_quotes}
\title{Removes quotes from data.frame columns}
\usage{
remove_quotes(text)
}
\arguments{
\item{text}{The text column to remove quotes from.}
}
\description{
Removes quotes from data.frame columns
}
