% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{replaceNAwithNULL}
\alias{replaceNAwithNULL}
\title{replaceNAwithNULL}
\usage{
replaceNAwithNULL(sql.code)
}
\arguments{
\item{sql.code}{code to replace NA with NULL}
}
\description{
Replace NA with NULL in sql statement
}
\author{
ken4rab
}
