% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{rm_quotes}
\alias{rm_quotes}
\title{rm_quotes}
\usage{
rm_quotes(text, quotes = "'")
}
\arguments{
\item{text}{The string to remove quotes from}

\item{quotes}{Quote characters}
}
\description{
Removes quotes from the String
}
