% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{rm_vector_quotes}
\alias{rm_vector_quotes}
\title{rm_vector_quotes}
\usage{
rm_vector_quotes(text.vector)
}
\arguments{
\item{text.vector}{The text vector to remove quotes from.}
}
\description{
Removes quotes from data.frame columns
}
\author{
ken4rab
}
