% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lib.R
\docType{data}
\name{RSQL.class}
\alias{RSQL.class}
\title{The class that provides the SQL functionality}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RSQL.class
}
\description{
The class that provides the SQL functionality
}
\keyword{datasets}
