% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsql_table.R
\docType{class}
\name{rsql_alias-class}
\alias{rsql_alias-class}
\title{Class containing a R reference to a SQL table reference with an alias}
\description{
Class containing a R reference to a SQL table reference with an alias
}
