% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsql_table.R
\name{rsql_alias}
\alias{rsql_alias}
\title{Create a \code{rsql_alias} to \code{object}}
\usage{
rsql_alias(object, alias)
}
\arguments{
\item{object}{what to reference}

\item{alias}{name to use as alias}
}
\description{
Create a \code{rsql_alias} to \code{object}
}
