% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{rsql_assign}
\alias{rsql_assign}
\title{(Re)Assign a function to construct a SQL function/operator}
\usage{
rsql_assign(...)
}
\arguments{
\item{...}{named list of functions to assign}
}
\description{
Functions and operators common to multiple implementations
should already be defined.  Users should only need to 
assign new functions when they wish to use non-standard syntax
for their DBMS.
}
