% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsql_functions.R
\name{rsql_colnames}
\alias{rsql_colnames}
\title{Generate the column names from a list of quoted rsql expressions}
\usage{
rsql_colnames(x)
}
\arguments{
\item{x}{a list a quoted rsql expressions}
}
\description{
Generate the column names from a list of quoted rsql expressions
}
