% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsql_functions.R
\name{rsql_colrefs}
\alias{rsql_colrefs}
\title{Create quoted column references from column names and a table name or alias}
\usage{
rsql_colrefs(colnames, alias = NULL)
}
\arguments{
\item{colnames}{column names}

\item{alias}{table name or alias}
}
\description{
Create quoted column references from column names and a table name or alias
}
\examples{

library(rsql)
rsql_colrefs(c("x","y"),"tab")

}
