% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsql_statement.R
\docType{class}
\name{rsql_compound_select_statement-class}
\alias{rsql_compound_select_statement-class}
\title{Class containing an R representation of a compound SQL select statement}
\description{
Class containing an R representation of a compound SQL select statement
}
