% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsql_statement.R
\name{rsql_compound_statement}
\alias{rsql_compound_statement}
\title{Explicitly create a \code{rsql_compound_select_statement} object}
\usage{
rsql_compound_statement(colrefs, rsql)
}
\description{
This should not need to be called by users.
}
