% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsql_functions.R
\name{rsql_select}
\alias{rsql_select}
\title{Construct a SQL query}
\usage{
rsql_select(data, expr)
}
\arguments{
\item{data}{object to be selected}

\item{...}{additional arguments to be selected

If \code{data} is a \code{\link{rsql_table}} then it and \code{...}
are evaluated as \code{from(data,list(...))}.  In addition to adding the 
resulting elements to the list of expressions to be selected,
\code{from(data)} is added to the statement as a quoted 
\code{\link{rsql}} expression.

Otherwise \code{data} and any subsequent arguments are treated as
quoted R expressions as created by \code{\link{rsql}} and added
to the list of things to be selected.  In this case, no from clause
is added to the statement.

In addition to being called directly, \code{rsql_select} is used
to add select expressions using \code{$select} on \code{rsql_table}
and \code{rsql_select_statement} objects.  See examples below.}
}
\description{
Constructs a basic SQL query from arguments
}
\examples{
library(rsql)
tab = rsql_table("tab",c('x','y'))
to_sql(rsql_select(tab))
to_sql(rsql_select(tab,x))
to_sql(rsql_select(from(tab,x)))
to_sql(tab$select())
to_sql(tab$select(x))

}
