% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsql_statement.R
\docType{class}
\name{rsql_select_statement-class}
\alias{rsql_select_statement-class}
\title{Class containing an R representation of a SQL statement}
\description{
Class containing an R representation of a SQL statement
}
