% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsql_statement.R
\name{rsql_select_statement}
\alias{rsql_select_statement}
\title{Explicitly create a \code{rsql_select_statement} object}
\usage{
rsql_select_statement(select = list(), from = list(), optional = list())
}
\description{
This should not need to be called by users. Users should
instead use rsql_select.
}
