% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsql_functions.R
\name{rsql_substitute}
\alias{rsql_substitute}
\title{Evaluate any bquoted expressions and substitute actual column references.}
\usage{
rsql_substitute(args, colrefs = list())
}
\description{
Evaluate any bquoted expressions and substitute actual column references.
}
