% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsql_table.R
\name{rsql_table}
\alias{rsql_table}
\title{Create an rsql table reference.}
\usage{
rsql_table(X, colnames = NULL)
}
\arguments{
\item{X}{table name}

\item{colnames}{vector of column names}
}
\description{
Create an rsql table reference.
}
