% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bquote.R
\name{sql_bquote}
\alias{sql_bquote}
\title{Just like regular bquote, but unlists results}
\usage{
sql_bquote(expr, where = parent.frame())
}
\arguments{
\item{expr}{the quoted expression}

\item{where}{the environment in which to evaluate any expressions}
}
\description{
Very similar to the base:bquote, but when the expression evaluates to a list,
the results are inserted into the quoted expression.
}
\examples{
a= list(quote(b),quote(c))
bquote(quote(f(.(a)))) #f(list(b, c))
library(rsql)
sql_bquote(quote(f(.(a)))) #f(b, c)
}
