% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{sql_execute_get_insert}
\alias{sql_execute_get_insert}
\title{sql_execute_get_insert}
\usage{
sql_execute_get_insert(dbconn, sql_select, sql_insert, ...)
}
\arguments{
\item{dbconn}{The db connection}

\item{sql_select}{The SQL select query}

\item{sql_insert}{The SQL insert query}

\item{...}{other variables to considered.}
}
\description{
Executes the insert statement
}
\author{
ken4rab
}
