% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lib.R
\name{sql_execute_insert}
\alias{sql_execute_insert}
\title{Extecutes a statement on the database.}
\usage{
sql_execute_insert(sql_insert, dbconn = NULL, export = c("db", "df"))
}
\arguments{
\item{sql_insert}{The SQL String}

\item{dbconn}{The Database Connection to run the query against}

\item{export}{The export type (either 'db' or 'df')}
}
\description{
Extecutes a statement on the database.
}
