% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lib.R
\name{sql_execute_select}
\alias{sql_execute_select}
\title{sql_execute_select}
\usage{
sql_execute_select(id_procesamiento, sql_select, dbconn = NULL,
  debug_level = 0)
}
\arguments{
\item{id_procesamiento}{TEST}

\item{sql_select}{TEST}

\item{dbconn}{TEST}

\item{debug_level}{TEST}
}
\description{
sql_execute_select
}
