% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lib.R
\name{sql_execute_update}
\alias{sql_execute_update}
\alias{sql_execute_update}
\title{sql_execute_update}
\usage{
sql_execute_update(id_procesamiento, sql_insert, dbconn = NULL,
  export = c("db", "df"), debug_level = 0)

sql_execute_update(id_procesamiento, sql_insert, dbconn = NULL,
  export = c("db", "df"), debug_level = 0)
}
\arguments{
\item{id_procesamiento}{TEST}

\item{sql_insert}{TEST}

\item{dbconn}{TEST}

\item{export}{TEST}

\item{debug_level}{TEST}

\item{sql_update}{TEST}

\item{id_procesamiento}{TEST}

\item{export}{TEST}

\item{debug_level}{TEST}
}
\description{
sql_execute_update

sql_execute_update
}
