% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lib.R
\name{sql_gen_delete.bad}
\alias{sql_gen_delete.bad}
\title{sql_gen_delete.bad}
\usage{
sql_gen_delete.bad(table, where_fields, where_values = c())
}
\arguments{
\item{table}{TEST}

\item{where_fields}{TEST}

\item{where_values}{TEST}
}
\description{
sql_gen_delete.bad
}
