% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lib.R
\name{sql_gen_delete}
\alias{sql_gen_delete}
\title{sql_gen_delete}
\usage{
sql_gen_delete(table, where_fields = "", where_values = NULL)
}
\arguments{
\item{table}{TEST}

\item{where_fields}{TEST}

\item{where_values}{TEST}
}
\description{
sql_gen_delete
}
