% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lib.R
\name{sql_gen_insert}
\alias{sql_gen_insert}
\title{sql_gen_insert}
\usage{
sql_gen_insert(table, insert_fields, values = c(), debug_level = 0)
}
\arguments{
\item{table}{TEST}

\item{insert_fields}{TEST}

\item{values}{TEST}

\item{debug_level}{TEST}
}
\description{
sql_gen_insert
}
