% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lib.R
\name{sql_gen_joined_query}
\alias{sql_gen_joined_query}
\title{Generates a Joined Query}
\usage{
sql_gen_joined_query(dw_definition, recipe, indicator_fields)
}
\arguments{
\item{dw_definition}{TEST}

\item{recipe}{TEST}

\item{indicator_fields}{TEST}
}
\description{
Generates a Joined Query
}
