% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lib.R
\name{sql_gen_update}
\alias{sql_gen_update}
\title{sql_execute_update}
\usage{
sql_gen_update(table, update_fields, values, fields_id, values_id,
  debug_level = 0)
}
\arguments{
\item{table}{TEST}

\item{update_fields}{TEST}

\item{values}{TEST}

\item{fields_id}{TEST}

\item{values_id}{TEST}

\item{debug_level}{TEST}
}
\description{
sql_execute_update
}
