% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{sql_gen_where.new}
\alias{sql_gen_where.new}
\title{Generates a where statement to be used on a SQL statement.}
\usage{
sql_gen_where.new(where_clause_list)
}
\arguments{
\item{where_clause_list}{The fields used in the where section}
}
\description{
Generates a where statement to be used on a SQL statement.
}
