% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{sql_gen_where_list.new}
\alias{sql_gen_where_list.new}
\title{Generates a where list statement to be used on a SQL statement.}
\usage{
sql_gen_where_list.new(where_clause_df)
}
\arguments{
\item{where_clause_df}{The fields used in the where section}
}
\description{
Generates a where list statement to be used on a SQL statement.
}
