% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lib.R
\name{sql_gen_where_or}
\alias{sql_gen_where_or}
\title{sql_gen_where_or}
\usage{
sql_gen_where_or(where_fields, where_values)
}
\arguments{
\item{where_fields}{TEST}

\item{where_values}{TEST}
}
\description{
sql_gen_where_or
}
