% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{sql_join}
\alias{sql_join}
\title{Format SQL JOIN statement to select the arguments passed}
\usage{
sql_join(x, on, method = NULL)
}
\arguments{
\item{x}{table reference to join}

\item{method}{type of join to perform, e.g. \code{"LEFT OUTER"}}

\item{condition}{condition on which to join}
}
\description{
Converts arguments to SQL using \code{to_sql} before formatting
them.
}
