% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{sql_retrieve}
\alias{sql_retrieve}
\title{Retrieves Statement}
\usage{
sql_retrieve(
  table,
  fields_uk = names(values_uk),
  values_uk,
  fields = names(values),
  values = NULL,
  field_id = "id",
  dbconn = NULL
)
}
\arguments{
\item{table}{The table}

\item{fields_uk}{The fields unique key}

\item{values_uk}{The values unique key}

\item{fields}{The fields (Not used. Included for compatibility)}

\item{values}{The values (Not used. Included for compatibility)}

\item{field_id}{The field of the serial id}

\item{dbconn}{The database connection}
}
\description{
Retrieves Statement
}
