% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lib.R
\name{sql_retrieve_insert}
\alias{sql_retrieve_insert}
\title{sql_retrieve_insert}
\usage{
sql_retrieve_insert(id_procesamiento, table, fields_id, values_id,
  fields = NULL, values = NULL, dbconn = NULL)
}
\arguments{
\item{id_procesamiento}{TEST}

\item{table}{TEST}

\item{fields_id}{TEST}

\item{fields}{TEST}

\item{values}{TEST}

\item{dbconn}{TEST}
}
\description{
sql_retrieve_insert
}
