% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{sql_select}
\alias{sql_select}
\title{Format SQL SELECT statement to select the arguments passed}
\usage{
sql_select(...)
}
\arguments{
\item{...}{named list of expressions to select}
}
\description{
Converts arguments to SQL using \code{to_sql} before formatting
them using \code{names} as aliases
}
