% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{sql_union}
\alias{sql_union}
\title{Format SQL UNION statement to select the arguments passed}
\usage{
sql_union(e1, e2, ..., all = TRUE)
}
\arguments{
\item{e1}{table reference to union}

\item{e2}{table reference to union}

\item{all}{logical indication whether to union all}
}
\description{
Converts arguments to SQL using \code{to_sql} before formatting
them.
}
