% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{stuff_df_quoted}
\alias{stuff_df_quoted}
\title{stuff quote characters in quoted or not quoted df for DSL or DML operations}
\usage{
stuff_df_quoted(text.df)
}
\arguments{
\item{text.df}{Data Frame with corresponding values and fields as colnames}
}
\description{
stuff quote characters in quoted or not quoted df for DSL or DML operations
}
