% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{times_to_utc}
\alias{times_to_utc}
\title{times_to_utc}
\usage{
times_to_utc(values.df)
}
\arguments{
\item{values.df}{Data Frame with corresponding values and fields as colnames}
}
\description{
This function converts POSIXct columns to UTC for inserting in a database
}
\author{
ken4rab
}
