% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0methods.R, R/to_sql.R
\docType{methods}
\name{to_sql}
\alias{to_sql}
\alias{to_sql}
\alias{to_sql.default}
\alias{to_sql.list}
\alias{to_sql.expression}
\alias{to_sql.quoted}
\alias{to_sql.name}
\alias{to_sql.function}
\alias{to_sql.NULL}
\alias{to_sql,NULL-method}
\alias{to_sql.call}
\alias{to_sql.(}
\alias{to_sql.if}
\title{Format an rsql object to SQL}
\usage{
to_sql(x)

to_sql(x)

to_sql.default(x)

to_sql.list(x)

to_sql.expression(x)

to_sql.quoted(x)

to_sql.name(x)

to_sql.function(x)

to_sql.NULL(x)

\S4method{to_sql}{`NULL`}(x)

to_sql.call(x)

"to_sql.("(x)

to_sql.if(x)
}
\description{
Format an rsql object to SQL

Format an rsql object to SQL
}
