% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lib.R
\name{trim.leading}
\alias{trim.leading}
\title{trim.leading}
\usage{
trim.leading(x)
}
\arguments{
\item{x}{TEST}

\item{update_fields}{TEST}

\item{values}{TEST}

\item{fields_id}{TEST}

\item{values_id}{TEST}

\item{debug_level}{TEST}
}
\description{
Returns string w/o leading whitespace
}
