% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{trim}
\alias{trim}
\title{Returns string w/o leading or trailing whitespace}
\usage{
trim(x)
}
\arguments{
\item{x}{The string}
}
\description{
Returns string w/o leading or trailing whitespace
}
\author{
ken4rab
}
