% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{unary_operator_to_sql}
\alias{unary_operator_to_sql}
\title{Format a unary operator with name \code{x}}
\usage{
unary_operator_to_sql(x, prefix = TRUE)
}
\arguments{
\item{x}{operator name}
}
\description{
Function created converts arguments \code{...} to SQL using
\code{to_sql} before formatting them either before or after
name \code{x}, depending on \code{prefix}.
}
