% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lib.R
\name{execute_get_insert}
\alias{execute_get_insert}
\title{Executes the insert statement}
\usage{
execute_get_insert(sql_select, sql_insert, ...)
}
\arguments{
\item{sql_select}{The SQL select query}

\item{sql_insert}{The SQL insert query}
}
\description{
Executes the insert statement
}
