% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{is.POSIXct}
\alias{is.POSIXct}
\title{is.POSIXct}
\usage{
is.POSIXct(x)
}
\arguments{
\item{x}{value to be checked as a POSIXct}
}
\description{
This function returns true if x is a POSIXct object type
}
\author{
ken4rab
}
