% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{re_quote}
\alias{re_quote}
\title{This functions remove original quotes and sets validated quotes for corresponding db.
If it had no quotes, will only put corresponding quotes symbols}
\usage{
re_quote(text, quotes = "'")
}
\arguments{
\item{text}{The string}

\item{quotes}{The quotes}
}
\description{
This functions remove original quotes and sets validated quotes for corresponding db.
If it had no quotes, will only put corresponding quotes symbols
}
