% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{rename_col}
\alias{rename_col}
\title{rename_col}
\usage{
rename_col(df, name, replace_name)
}
\arguments{
\item{df}{The date.frame}

\item{name}{The name of the column}

\item{replace_name}{The new name of the column}
}
\description{
renames a column on a data.frame
}
