% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{RSQL.class}
\alias{RSQL.class}
\title{The class that provides the SQL functionality.}
\description{
This class is intended to simplify SQL commands.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RSQL.class$new()}}
\item \href{#method-checkEntityName}{\code{RSQL.class$checkEntityName()}}
\item \href{#method-gen_select}{\code{RSQL.class$gen_select()}}
\item \href{#method-gen_insert}{\code{RSQL.class$gen_insert()}}
\item \href{#method-gen_update}{\code{RSQL.class$gen_update()}}
\item \href{#method-gen_delete}{\code{RSQL.class$gen_delete()}}
\item \href{#method-execute_select}{\code{RSQL.class$execute_select()}}
\item \href{#method-execute_update}{\code{RSQL.class$execute_update()}}
\item \href{#method-execute_insert}{\code{RSQL.class$execute_insert()}}
\item \href{#method-execute_delete}{\code{RSQL.class$execute_delete()}}
\item \href{#method-retrieve_insert}{\code{RSQL.class$retrieve_insert()}}
\item \href{#method-disconnect}{\code{RSQL.class$disconnect()}}
\item \href{#method-clone}{\code{RSQL.class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RSQL.class$new(
  drv,
  dbname,
  user = NULL,
  password = NULL,
  host = NULL,
  port = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkEntityName"></a>}}
\subsection{Method \code{checkEntityName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RSQL.class$checkEntityName(fields)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_select"></a>}}
\subsection{Method \code{gen_select()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RSQL.class$gen_select(
  select_fields,
  table,
  where_fields = names(where_values),
  where_values = NULL,
  group_by = c(),
  order_by = c(),
  top = 0,
  distinct = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_insert"></a>}}
\subsection{Method \code{gen_insert()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RSQL.class$gen_insert(table, values_df, insert_fields = names(values_df))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values_df}}{The values to insert. Must be defined as data.frame of values
Composite function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_update"></a>}}
\subsection{Method \code{gen_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RSQL.class$gen_update(
  table,
  update_fields = names(values),
  values,
  where_fields = names(where_values),
  where_values = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_delete"></a>}}
\subsection{Method \code{gen_delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RSQL.class$gen_delete(
  table,
  where_fields = names(where_values),
  where_values = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute_select"></a>}}
\subsection{Method \code{execute_select()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RSQL.class$execute_select(sql_select)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute_update"></a>}}
\subsection{Method \code{execute_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RSQL.class$execute_update(sql_update)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute_insert"></a>}}
\subsection{Method \code{execute_insert()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RSQL.class$execute_insert(sql_insert, export = "db")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute_delete"></a>}}
\subsection{Method \code{execute_delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RSQL.class$execute_delete(sql_delete)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-retrieve_insert"></a>}}
\subsection{Method \code{retrieve_insert()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RSQL.class$retrieve_insert(
  table,
  fields_uk = names(values_uk),
  values_uk,
  fields = names(values),
  values,
  field_id = "id"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-disconnect"></a>}}
\subsection{Method \code{disconnect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RSQL.class$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RSQL.class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
