% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R, R/zzz.R
\docType{package}
\name{rsql}
\alias{rsql}
\alias{.onLoad}
\title{Produces a RSQL object}
\usage{
rsql(drv, dbname, user = NULL, password = NULL, host = NULL, port = NULL)

.onLoad(libname, pkgname)
}
\arguments{
\item{drv}{Driver name}

\item{dbname}{Database name}

\item{user}{Database user name}

\item{password}{Database password}

\item{host}{Database host}

\item{port}{Database port}

\item{libname}{Library name}

\item{pkgname}{Package name}
}
\description{
A package to work with SQL datasources in a simple manner
}
\author{
Leonardo Javier Belen <leobelen@gmail.com>, Alejandro Baranek <alejandrobaranek@gmail.com>
Executes code while loading the package.
}
