% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lib.R, R/zzz.R
\docType{package}
\name{rsql}
\alias{rsql}
\alias{rsql}
\alias{.onLoad}
\alias{rsql-package}
\title{Produces a RSQL object}
\usage{
rsql(drv, dbname, user = NULL, password = NULL, host = NULL,
  port = NULL)

.onLoad(libname, pkgname)
}
\arguments{
\item{drv}{The DBI compatible driver}

\item{dbname}{The database name}

\item{user}{The database user}

\item{password}{The database password}

\item{host}{The database host}

\item{port}{The database port}
}
\description{
Produces a RSQL object

A package to work with SQL datasources in a simple manner
}
\author{
Leonardo Javier Belen <leobelen@gmail.com>, Alejandro Baranek <alejandrobaranek@gmail.com>
}
