% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lib.R
\name{sql_execute_delete}
\alias{sql_execute_delete}
\title{sql_execute_delete}
\usage{
sql_execute_delete(id_procesamiento, sql_delete, dbconn = NULL)
}
\arguments{
\item{id_procesamiento}{TEST}

\item{sql_insert}{TEST}
}
\description{
sql_execute_delete
}
