% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{sql_execute_delete}
\alias{sql_execute_delete}
\title{sql_execute_delete}
\usage{
sql_execute_delete(sql_delete, dbconn = NULL)
}
\arguments{
\item{sql_delete}{The delete SQL}

\item{dbconn}{The Database Connection to run the query against}
}
\description{
Executes a delete on the Database
}
