% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lib.R
\name{sql_execute_insert}
\alias{sql_execute_insert}
\title{sql_execute_insert}
\usage{
sql_execute_insert(id_procesamiento, sql_insert, dbconn = NULL,
  export = c("db", "df"), debug_level = 0)
}
\arguments{
\item{id_procesamiento}{TEST}

\item{sql_insert}{TEST}

\item{export}{TEST}

\item{debug_level}{TEST}
}
\description{
sql_execute_insert
}
