% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{sql_execute_insert}
\alias{sql_execute_insert}
\title{Executes a statement on the database.}
\usage{
sql_execute_insert(sql_insert, dbconn)
}
\arguments{
\item{sql_insert}{The SQL String}

\item{dbconn}{The Database Connection to run the query against}
}
\description{
Executes a statement on the database.
}
\author{
ken4rab
}
