% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{sql_execute_select}
\alias{sql_execute_select}
\title{sql_execute_select}
\usage{
sql_execute_select(sql_select, dbconn = NULL)
}
\arguments{
\item{sql_select}{The delete SQL}

\item{dbconn}{The Database Connection to run the query against}
}
\description{
Executes a select on the database
}
\author{
ken4rab
}
