% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{sql_execute_update}
\alias{sql_execute_update}
\alias{sql_execute_update}
\title{Executes an update on the database}
\usage{
sql_execute_update(sql_insert, dbconn = NULL, export = c("db", "df"))

sql_execute_update(sql_insert, dbconn = NULL, export = c("db", "df"))
}
\arguments{
\item{sql_insert}{The insert statement}

\item{dbconn}{The Database Connection to run the query against}

\item{export}{The export type (either 'db' or 'df')}

\item{sql_update}{The update SQL}

\item{dbconn}{the Database Connection}

\item{export}{The export method}
}
\description{
Executes an update on the database

Executes an update
}
