% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{sql_execute_update}
\alias{sql_execute_update}
\title{Executes an update on the database}
\usage{
sql_execute_update(sql_update, dbconn = NULL)
}
\arguments{
\item{sql_update}{The update SQL}

\item{dbconn}{The Database Connection to run the query against}
}
\description{
Executes an update on the database
}
\author{
ken4rab
}
