% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{sql_gen_delete}
\alias{sql_gen_delete}
\title{Generates a Delete Statement}
\usage{
sql_gen_delete(table, where_fields = names(where_values), where_values = NULL)
}
\arguments{
\item{table}{The table from which the delete statement will be generated}

\item{where_fields}{The fields used in the where section}

\item{where_values}{The values used in the where section}
}
\description{
Generates a Delete Statement
}
