% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{sql_gen_insert}
\alias{sql_gen_insert}
\title{Generates an insert statement.}
\usage{
sql_gen_insert(table, insert_fields, values = c())
}
\arguments{
\item{table}{The table to be affected}

\item{insert_fields}{The fields to insert}

\item{values}{The values to insert}

\item{where_fields}{The fields used in the where section}

\item{where_values}{The values used in the where section}
}
\description{
Generates an insert statement.
}
