% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{sql_gen_insert}
\alias{sql_gen_insert}
\title{Generates an insert statement.}
\usage{
sql_gen_insert(table, values_df, insert_fields = names(values_df))
}
\arguments{
\item{table}{The table to be affected}

\item{values_df}{The values to insert. Must be defined as data.frame of values}

\item{insert_fields}{The fields to insert}
}
\description{
Generates an insert statement.
}
