% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{sql_gen_select}
\alias{sql_gen_select}
\title{Generates a Select Statement}
\usage{
sql_gen_select(select_fields, table, where_fields = NULL,
  where_values = NULL, group_by = c(), order_by = c(), top = 0)
}
\arguments{
\item{select_fields}{The fields to be selected}

\item{table}{The table to be used in the select}

\item{where_fields}{The fields used in the where section}

\item{where_values}{The values used in the where section}

\item{group_by}{Group by fields}

\item{order_by}{Order by fields}

\item{top}{Retrieve top records}
}
\description{
Generates a Select Statement
}
