% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{sql_gen_update}
\alias{sql_gen_update}
\title{Generates an update statement}
\usage{
sql_gen_update(table, update_fields, values, fields_id, values_id)
}
\arguments{
\item{table}{The table to update}

\item{update_fields}{The fields to update}

\item{values}{The values to update}

\item{fields_id}{The fields id}

\item{values_id}{The values id}
}
\description{
Generates an update statement
}
