% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{sql_gen_update}
\alias{sql_gen_update}
\title{sql_gen_update}
\usage{
sql_gen_update(
  table,
  update_fields = names(values),
  values,
  where_fields = names(where_values),
  where_values
)
}
\arguments{
\item{table}{The table to update}

\item{update_fields}{The fields to update}

\item{values}{The values to update}

\item{where_fields}{The fields for where statement}

\item{where_values}{The values for where statement}
}
\description{
Generates an update statement
}
\author{
ken4rab
}
