% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-lib.R
\name{sql_gen_where}
\alias{sql_gen_where}
\title{sql_gen_where}
\usage{
sql_gen_where(where_fields = names(where_values), where_values)
}
\arguments{
\item{where_fields}{The fields used in the where section}

\item{where_values}{The values used in the where section}
}
\description{
Generates a where statement to be used on a SQL statement.
}
