% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lib.R
\name{sql_retrieve_insert}
\alias{sql_retrieve_insert}
\title{Retrieves an insert Statement}
\usage{
sql_retrieve_insert(table, fields_id, values_id, fields = NULL,
  values = NULL, dbconn = NULL)
}
\arguments{
\item{table}{The table}

\item{fields_id}{The fields ID}

\item{values_id}{The value ID}

\item{fields}{The fields}

\item{values}{The values}

\item{dbconn}{The database connection}
}
\description{
Retrieves an insert Statement
}
